Pull down cross compile dependencies on jenkins
Jenkins doesn't really know when pulling down artifacts whether it is
cross compiling.  We added TR logic to only pull down the cross compile
toolchain when we are cross compiling, but we still need to pull down
cross compilied dependencies to build against.  Unlike the toolchain
though these are tiny so it shouldn't hurt just to pull them down.
There isn't any way to mark a dependency as optional though so we
have to explicitly specify which platforms we expect to have a cross
compile artifact, which makes this especially hacky.

This also updates a few things in the gitignore