Cleanup docs, whitespace and language
The jsontoolpath spec was ridiculously out of date, including listing commands
which have never been implemented and even referencing a completely different
proposed implementation of tags than what we ended up using.  It also tried
to document a variable substitution / expression evaluation system that was
never used and was ripped out in the parser refactor.

I also added an alternate form of comment to the spec here.  The parser has
actually supported this since the refactor, in that it ignores basically
everything that it does not know about.  I added a few lines here and there
recommending that parser implementations do this for forward compatibility
reasons, but I think it is good to put this specifically in the spec to
try and standardize comments on something less insane than the 2.0 comment
format.