Special handling for 'sketch' bot_type...
BW-5082
http://makerbot.atlassian.net/browse/BW-5082

because it doesn't follow the typical format of [name]_[pid]

Also we make sure that we do not crash on any unexpected formatting of this
field.  In lieu of catching exceptions thrown by stoi, we simply use strtol,
and we just consider 0 to no longer be a valid PID to simplify the logic for
ensuring that 9999 is returned for an invalid bot_type string.